+++ NEWS.md
@@ -5,6 +5,8 @@
     object to a lower resolution, using the specific calendar of the temporal
     dimension and returns an array with the summarised data.
 -   Method `CFData$data.table()` exports a data object to a `data.table`.
+-   `CFVariable` and `CFData` classes now have `time()` method to retrieve the
+    `CFTime` instance of a "time" axis, if present.
 -   New attributes can be defined on any object that supports attributes, or
     deleted.